requirements.txt
fastapi
uvicorn
transformers
torch
pyyaml
pillow
python-multipart
pydub
openai-whisper